Alex/ctc pack rebase (#9)
* Try adding collapse across dimensions

* valid_pct is now an int32 like the good ol' days

* Fix transcript test

+++ loader/src/provider_audio_transcriber.cpp
@@ -35,7 +35,7 @@ audio_transcriber::audio_transcriber(nlohmann::json js) :
     shape_type trans_length({1, 1}, output_type("uint32_t"));
     oshapes.push_back(trans_length);
 
-    shape_type valid_pct({1, 1}, output_type("float"));
+    shape_type valid_pct({1, 1}, output_type("uint32_t"));
     oshapes.push_back(valid_pct);
 }
 
@@ -64,7 +64,7 @@ void audio_transcriber::provide(int idx, buffer_in_array& in_buf, buffer_out_arr
 
     // Get the length of each audio record as a percentage of
     // maximum utterance length
-    float valid_pct = 100 * (float)audio_dec->valid_frames / (float)audio_config.time_steps;
+    uint32_t valid_pct = 100 * audio_dec->valid_frames / audio_config.time_steps;
 
     pack(valid_out, valid_pct);
 }
